+++ locals.tf
@@ -5,10 +5,12 @@ locals {
 
   hostname = coalesce(var.custom_vm_hostname, local.default_basename)
 
+  ssh_public_key  = coalesce(var.ssh_public_key, tls_private_key.ssh.public_key_openssh)
+  ssh_private_key = coalesce(var.ssh_private_key, tls_private_key.ssh.private_key_pem)
+
   bastion_tags = {
     env    = var.environment
     stack  = var.stack
     module = "bastion"
   }
 }
-