+++ r-keys.tf
@@ -1,0 +1,3 @@
+resource "tls_private_key" "ssh" {
+  algorithm = "RSA"
+}