+++ versions.tf
@@ -17,5 +17,9 @@ terraform {
       source  = "hashicorp/template"
       version = ">= 2.0"
     }
+    tls = {
+      source  = "hashicorp/tls"
+      version = ">= 3.0"
+    }
   }
 }