+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Fixed
+  * AZ-607: Update example usage and wait until the bastion VM is reachable before executing the Ansible playbook
+
 # v4.3.0 - 2021-10-19
 
 Breaking